# docs/swagger.yaml
definitions:
  fail.Response:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  ? github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_profile_dto_request_body.ConfirmAvatarUploading
  : properties:
      avatarId:
        type: string
    type: object
...
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_profile_dto_request_body.SetDescription:
    properties:
      description:
        type: string
    type: object
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_profile_dto_request_body.SetName:
    properties:
      firstName:
        type: string
      lastName:
        type: string
    type: object
...
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_common_body.CookingItem:
    properties:
      id:
        type: string
      pictures:
        items:
          type: string
        type: array
      recipeId:
        type: string
      text:
        type: string
      time:
        type: integer
      type:
        type: string
    type: object
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_common_body.IngredientItem:
    properties:
      amount:
        type: integer
      id:
        type: string
      recipeId:
        type: string
      text:
        type: string
      type:
        type: string
      unit:
        type: string
    type: object
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_common_body.Macronutrients:
    properties:
      carbohydrates:
        type: integer
      fats:
        type: integer
      protein:
        type: integer
    type: object
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_request_body.GetRecipes:
    properties:
      authorId:
        type: string
      lastCalories:
        type: integer
      lastCreationTimestamp:
        type: string
      lastRating:
        type: number
      lastRecipeId:
        type: string
      lastTime:
        type: integer
      lastUpdateTimestamp:
        type: string
      lastVotes:
        type: integer
      maxCalories:
        type: integer
      maxRating:
        type: integer
      maxServings:
        type: integer
      maxTime:
        type: integer
      minCalories:
        type: integer
      minRating:
        type: integer
      minServings:
        type: integer
      minTime:
        type: integer
      owned:
        type: boolean
      pageSize:
        type: integer
      recipeLanguages:
        items:
          type: string
        type: array
      saved:
        type: boolean
      search:
        type: string
      sorting:
        type: string
      userLanguage:
        type: string
    type: object
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_request_body.RecipeInput:
    properties:
      calories:
        type: integer
      cooking:
        items:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_common_body.CookingItem'
        type: array
      description:
        type: string
      encrypted:
        type: boolean
      ingredients:
        items:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_common_body.IngredientItem'
        type: array
      language:
        type: string
      macronutrients:
        $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_common_body.Macronutrients'
      name:
        type: string
      recipeId:
        type: string
      servings:
        type: integer
      tags:
        items:
          type: string
        type: array
      time:
        type: integer
      version:
        type: integer
      visibility:
        type: string
    required:
    - cooking
    - ingredients
    - name
    type: object
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.Category:
    properties:
      emoji:
        type: string
      name:
        type: string
    type: object
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.CreateRecipe:
    properties:
      recipeId:
        type: string
      version:
        type: integer
    type: object
  ? github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.GetRecipeBookResponse
  : properties:
      categories:
        items:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.Category'
        type: array
      recipes:
        items:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.RecipeState'
        type: array
      tags:
        additionalProperties:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.Tag'
        type: object
    type: object
  ? github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.GetRecipeResponse
  : properties:
      categories:
        additionalProperties:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.Category'
        type: object
      recipe:
        $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.Recipe'
      tags:
        additionalProperties:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.Tag'
        type: object
    type: object
  ? github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.GetRecipesResponse
  : properties:
      categories:
        additionalProperties:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.Category'
        type: object
      recipes:
        items:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.RecipeInfo'
        type: array
      tags:
        additionalProperties:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.Tag'
        type: object
    type: object
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.Recipe:
    properties:
      calories:
        type: integer
      categories:
        items:
          type: string
        type: array
      cooking:
        items:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_common_body.CookingItem'
        type: array
      creationTimestamp:
        type: string
      description:
        type: string
      encrypted:
        type: boolean
      favourite:
        type: boolean
      ingredients:
        items:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_common_body.IngredientItem'
        type: array
      language:
        type: string
      macronutrients:
        $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_common_body.Macronutrients'
      name:
        type: string
      owned:
        type: boolean
      ownerAvatar:
        type: string
      ownerId:
        type: string
      ownerName:
        type: string
      preview:
        type: string
      rating:
        type: number
      recipeId:
        type: string
      saved:
        type: boolean
      score:
        type: integer
      servings:
        type: integer
      tags:
        items:
          type: string
        type: array
      time:
        type: integer
      updateTimestamp:
        type: string
      version:
        type: integer
      visibility:
        type: string
      votes:
        type: integer
    type: object
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.RecipeInfo:
    properties:
      calories:
        type: integer
      categories:
        items:
          type: string
        type: array
      creationTimestamp:
        type: string
      encrypted:
        type: boolean
      favourite:
        type: boolean
      language:
        type: string
      name:
        type: string
      owned:
        type: boolean
      ownerAvatar:
        type: string
      ownerId:
        type: string
      ownerName:
        type: string
      preview:
        type: string
      rating:
        type: number
      recipeId:
        type: string
      saved:
        type: boolean
      score:
        type: integer
      servings:
        type: integer
      tags:
        items:
          type: string
        type: array
      time:
        type: integer
      updateTimestamp:
        type: string
      version:
        type: integer
      visibility:
        type: string
      votes:
        type: integer
    type: object
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.RecipeState:
    properties:
      categories:
        items:
          type: string
        type: array
      favourite:
        type: boolean
      ownerAvatar:
        type: string
      ownerName:
        type: string
      preview:
        type: string
      rating:
        type: number
      recipeId:
        type: string
      score:
        type: integer
      version:
        type: integer
      votes:
        type: integer
    type: object
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.Tag:
    properties:
      emoji:
        type: string
      name:
        type: string
    type: object
...
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.LinkBody:
    properties:
      link:
        type: string
    type: object
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/profile/avatar:
    delete:
      consumes:
      - application/json
      description: Delete avatar
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete avatar
      tags:
      - user
      - profile
    post:
      consumes:
      - application/json
      description: Delete avatar
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.LinkBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete avatar
      tags:
      - user
      - profile
    put:
      consumes:
      - application/json
      description: Confirm avatar uploading
      parameters:
      - description: Avatar ID
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_profile_dto_request_body.ConfirmAvatarUploading'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Confirm avatar uploading
      tags:
      - user
      - profile
...
  /v1/profile/description:
    put:
      consumes:
      - application/json
      description: Set profile description
      parameters:
      - description: Description
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_profile_dto_request_body.SetDescription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Set description
      tags:
      - user
      - profile
  /v1/profile/name:
    put:
      consumes:
      - application/json
      description: Set profile name
      parameters:
      - description: Name
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_profile_dto_request_body.SetName'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Set name
      tags:
      - user
      - profile
...
  /v1/recipes:
    get:
      consumes:
      - application/json
      description: Get recipes by query
      parameters:
      - description: Query
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_request_body.GetRecipes'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.GetRecipesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Get recipes
      tags:
      - recipe
    post:
      consumes:
      - application/json
      description: Create recipe
      parameters:
      - description: Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_request_body.RecipeInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.CreateRecipe'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Create recipe
      tags:
      - recipe
...
  /v1/recipes/book:
    get:
      consumes:
      - application/json
      description: Get user recipe book
      parameters:
      - description: User language code
        in: query
        name: user_language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.GetRecipeBookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Get recipe book
      tags:
      - recipe
...
  /v1/recipes/random:
    get:
      consumes:
      - application/json
      description: Get random recipe
      parameters:
      - collectionFormat: csv
        description: Recipe language codes
        in: query
        items:
          type: string
        name: recipe_language
        type: array
      - description: User language code
        in: query
        name: user_language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_handler_v1_recipe_dto_response_body.GetRecipeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Get random recipe
      tags:
      - recipe
...
swagger: "2.0"