# internal/transport/http/handler/v1/recipe/dto/common_body/ingredient.go
package common_body

type IngredientItem struct {
	Id       string  `json:"id"`
	Text     *string `json:"text,omitempty"`
	Type     string  `json:"type"`
	Amount   *int32  `json:"amount,omitempty"`
	Unit     *string `json:"unit,omitempty"`
	RecipeId *string `json:"recipeId,omitempty"`
}